refactor: lottoMatch도출
