--- src/main/java/step3/domain/lotto/LottoStatistics.java
@@ -1,10 +1,0 @@
-package step3.domain.lotto;
-
-public class LottoStatistics {
-
-    public static Profit calculateLottoProfit(LottoMatch lottoMatch, int lottoCount) {
-        return Profit.calcLottoProfit(lottoMatch.sumMoney(), lottoCount);
-    }
-
-
-}
