+++ src/test/java/step3/domain/lotto/LottoTest.java
@@ -1,0 +1,7 @@
+package step3.domain.lotto;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class LottoTest {
+
+}
